refactor(ApiApplication): 이름 변경 및 위치 이동
com.cosi.api.ApiApplication -> com.cosi.CosiApplication

관련 이슈
#1
