+++ api/src/main/java/com/cosi/CosiApplication.java
@@ -1,13 +1,13 @@
-package com.cosi.api;
+package com.cosi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ApiApplication {
+public class CosiApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiApplication.class, args);
+		SpringApplication.run(CosiApplication.class, args);
 	}
 
 }
