+++ api/src/test/java/com/cosi/api/CosiApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiApplicationTests {
+class CosiApplicationTests {
 
 	@Test
 	void contextLoads() {
